move constants file to the src folder and change its casing to camel case
